requirements.txt
requests
python-dotenv
pyqt6